# apidog-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Weblog API",
    "description": "NestJS weblog API using TypeORM and PostgreSQL.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Register" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered user and access token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Login" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Access token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Current user profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/categories": {
      "post": {
        "tags": ["Categories"],
        "summary": "Create category",
        "description": "Slug is optional. If provided, it is normalized. If omitted, it is generated from name.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCategory" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created category",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Categories"],
        "summary": "List categories",
        "responses": {
          "200": {
            "description": "Categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Category" }
                }
              }
            }
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get category by ID",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Category",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Categories"],
        "summary": "Update category",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCategory" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated category",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "Delete category",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/tags": {
      "post": {
        "tags": ["Tags"],
        "summary": "Create tag",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTag" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created tag",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Tag" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Tags"],
        "summary": "List tags",
        "responses": {
          "200": {
            "description": "Tags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Tag" }
                }
              }
            }
          }
        }
      }
    },
    "/tags/{id}": {
      "get": {
        "tags": ["Tags"],
        "summary": "Get tag by ID",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Tag",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Tag" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Tags"],
        "summary": "Update tag",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTag" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated tag",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Tag" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Tags"],
        "summary": "Delete tag",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/posts": {
      "post": {
        "tags": ["Posts"],
        "summary": "Create post",
        "description": "Requires a bearer token. The authenticated user becomes the post author; authorId in the body is ignored.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePost" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created post",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Post" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Posts"],
        "summary": "List posts",
        "parameters": [
          { "$ref": "#/components/parameters/PageQuery" },
          { "$ref": "#/components/parameters/LimitQuery" },
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "categoryId",
            "in": "query",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "tagSlug",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/PostStatus" }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated posts",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedPosts" }
              }
            }
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Get post by ID",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Post",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Post" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Posts"],
        "summary": "Update post",
        "description": "Only the post author can update the post.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePost" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated post",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Post" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Posts"],
        "summary": "Delete post",
        "description": "Only the post author can delete the post.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/posts/slug/{slug}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Get post by slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Post",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Post" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/comments": {
      "post": {
        "tags": ["Comments"],
        "summary": "Create comment",
        "description": "Requires a bearer token. The authenticated user becomes the comment author; authorId in the body is ignored.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateComment" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created comment",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Comment" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Comments"],
        "summary": "List comments",
        "responses": {
          "200": {
            "description": "Comments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Comment" }
                }
              }
            }
          }
        }
      }
    },
    "/comments/{id}": {
      "get": {
        "tags": ["Comments"],
        "summary": "Get comment by ID",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Comment",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Comment" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Comments"],
        "summary": "Update comment",
        "description": "Only the comment author can update the comment.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateComment" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated comment",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Comment" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Comments"],
        "summary": "Delete comment",
        "description": "Only the comment author can delete the comment.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "PageQuery": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "LimitQuery": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
...
      "PostStatus": {
        "type": "string",
        "enum": ["draft", "published"],
        "example": "published"
      },
      "Register": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 120 },
          "email": { "type": "string", "format": "email", "maxLength": 180 },
          "password": { "type": "string", "minLength": 8, "maxLength": 72 },
          "bio": { "type": "string", "maxLength": 500 }
        },
        "example": {
          "name": "Ali Reza",
          "email": "ali@example.com",
          "password": "password123",
          "bio": "Backend developer"
        }
      },
      "Login": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "maxLength": 180 },
          "password": { "type": "string", "minLength": 8, "maxLength": 72 }
        },
        "example": {
          "email": "ali@example.com",
          "password": "password123"
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "tokenType": { "type": "string", "example": "Bearer" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
      "CreateUser": {
        "type": "object",
        "required": ["name", "email"],
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 120 },
          "email": { "type": "string", "format": "email", "maxLength": 180 },
          "bio": { "type": "string", "maxLength": 500 }
        },
        "example": {
          "name": "Ali Reza",
          "email": "ali@example.com",
          "bio": "Backend developer"
        }
      },
...
      "User": {
        "allOf": [
          { "$ref": "#/components/schemas/CreateUser" },
          { "$ref": "#/components/schemas/Timestamps" },
          {
            "type": "object",
            "properties": {
              "id": { "type": "string", "format": "uuid" }
            }
          }
        ]
      },
      "CreateCategory": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 120 },
          "slug": {
            "type": "string",
            "minLength": 1,
            "maxLength": 140,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "description": { "type": "string", "maxLength": 500 }
        },
        "example": {
          "name": "Web Development",
          "description": "Frontend and backend posts"
        }
      },
      "UpdateCategory": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 120 },
          "slug": {
            "type": "string",
            "minLength": 1,
            "maxLength": 140,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "description": { "type": "string", "maxLength": 500 }
        }
      },
      "Category": {
        "allOf": [
          { "$ref": "#/components/schemas/CreateCategory" },
          { "$ref": "#/components/schemas/Timestamps" },
          {
            "type": "object",
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "slug": { "type": "string" }
            }
          }
        ]
      },
      "CreateTag": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 80 },
          "slug": { "type": "string", "maxLength": 100 }
        },
        "example": {
          "name": "NestJS"
        }
      },
      "UpdateTag": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 80 },
          "slug": { "type": "string", "maxLength": 100 }
        }
      },
      "Tag": {
        "allOf": [
          { "$ref": "#/components/schemas/CreateTag" },
          { "$ref": "#/components/schemas/Timestamps" },
          {
            "type": "object",
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "slug": { "type": "string" }
            }
          }
        ]
      },
      "CreatePost": {
        "type": "object",
        "required": ["title", "excerpt", "content", "categoryId"],
        "properties": {
          "title": { "type": "string", "minLength": 3, "maxLength": 180 },
          "slug": { "type": "string", "maxLength": 220 },
          "excerpt": { "type": "string", "minLength": 10, "maxLength": 300 },
          "content": { "type": "string", "minLength": 20 },
          "status": { "$ref": "#/components/schemas/PostStatus" },
          "authorId": { "type": "string", "format": "uuid" },
          "categoryId": { "type": "string", "format": "uuid" },
          "tagIds": {
            "type": "array",
            "uniqueItems": true,
            "items": { "type": "string", "format": "uuid" }
          }
        },
        "example": {
          "title": "First Post",
          "excerpt": "This is a short excerpt for the first post.",
          "content": "This is the full post content. It must be at least twenty characters long.",
          "status": "published",
          "authorId": "00000000-0000-4000-8000-000000000001",
          "categoryId": "00000000-0000-4000-8000-000000000002",
          "tagIds": ["00000000-0000-4000-8000-000000000003"]
        }
      },
      "UpdatePost": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 3, "maxLength": 180 },
          "slug": { "type": "string", "maxLength": 220 },
          "excerpt": { "type": "string", "minLength": 10, "maxLength": 300 },
          "content": { "type": "string", "minLength": 20 },
          "status": { "$ref": "#/components/schemas/PostStatus" },
          "authorId": { "type": "string", "format": "uuid" },
          "categoryId": { "type": "string", "format": "uuid" },
          "tagIds": {
            "type": "array",
            "uniqueItems": true,
            "items": { "type": "string", "format": "uuid" }
          }
        }
      },
      "Post": {
        "allOf": [
          { "$ref": "#/components/schemas/CreatePost" },
          { "$ref": "#/components/schemas/Timestamps" },
          {
            "type": "object",
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "publishedAt": { "type": "string", "format": "date-time", "nullable": true },
              "deletedAt": { "type": "string", "format": "date-time", "nullable": true },
              "author": { "$ref": "#/components/schemas/User" },
              "category": { "$ref": "#/components/schemas/Category" },
              "tags": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/Tag" }
              }
            }
          }
        ]
      },
      "PaginatedPosts": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Post" }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": { "type": "integer" },
              "limit": { "type": "integer" },
              "total": { "type": "integer" },
              "totalPages": { "type": "integer" }
            }
          }
        }
      },
      "CreateComment": {
        "type": "object",
        "required": ["authorName", "content", "postId"],
        "properties": {
          "authorName": { "type": "string", "minLength": 2, "maxLength": 120 },
          "authorEmail": { "type": "string", "format": "email", "maxLength": 180 },
          "content": { "type": "string", "minLength": 2 },
          "postId": { "type": "string", "format": "uuid" },
          "authorId": { "type": "string", "format": "uuid" }
        },
        "example": {
          "authorName": "Visitor",
          "authorEmail": "visitor@example.com",
          "content": "Nice post.",
          "postId": "00000000-0000-4000-8000-000000000004"
        }
      },
      "UpdateComment": {
        "type": "object",
        "properties": {
          "authorName": { "type": "string", "minLength": 2, "maxLength": 120 },
          "authorEmail": { "type": "string", "format": "email", "maxLength": 180 },
          "content": { "type": "string", "minLength": 2 }
        }
      },
      "Comment": {
        "allOf": [
          { "$ref": "#/components/schemas/CreateComment" },
          { "$ref": "#/components/schemas/Timestamps" },
          {
            "type": "object",
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "deletedAt": { "type": "string", "format": "date-time", "nullable": true },
              "post": { "$ref": "#/components/schemas/Post" },
              "author": {
                "allOf": [{ "$ref": "#/components/schemas/User" }],
                "nullable": true
              }
            }
          }
        ]
      },
      "Timestamps": {
        "type": "object",
        "properties": {
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "oneOf": [
              { "type": "string" },
              {
                "type": "array",
                "items": { "type": "string" }
              }
            ]
          },
          "error": { "type": "string" },
          "statusCode": { "type": "integer" }
        }
      }
    }
  }
}